Switch ExactExtract arguments (weighted->mean)
Since we do not provide any external weight files to ExactExtract for AORC forcings, we should just switch the ExactExtract argument (-s) from "weighted_mean" and "weighted_sum" to "mean" and "sum". This switch in syntax will stick to ExactExtract documentation and avoid confusing the user with this argument for regridding meteorological forcings.