Change sum -> mean for APCP_surface
Based on discussions with NextGen calibration teams, we've discovered that averaging APCP_surface values was the correction aggregation method based on the units being per-square meter. This signals that spatial aggregation is the appropriate method to average rain rate data in the framework, so ExactExtract regridding should also be accounting for spatial aggregation in all meteorological variables.