AORC BMI unit test for model engine team
Instructions to execute the AORC BMI module are located in the "README.md" file. The "requirements.txt" file highlight the required python packages needed to properly execute the AORC BMI module. Currently, this AORC BMI unit module dynamically downloads AORC data from the ERRDAP server during the model execution. Therefore, the user will need to test this script on a Linux environment that is capable of connecting to the server.