Delete regrid_ExactExtract.py
No use for this file moving forward for the NextGen project. This file was a simple python script to utilize a precompiled "ExactExtract" C++ executable to regrid AORC files for initial forcing provider testing in 2022. The NextGen Lumped Forcings Driver has now streamlined this process dynamically with a Python module containing ExactExtract bindings for AORC data and other meteorological models as well. Cleaning this file out to simplify the structure of the ngen-forcing directory.